+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 [Commits](https://github.com/thebigmunch/google-music-proto/compare/2.4.0...master)
 
+### Added
+
 * Ability to subscribe to public playlists using
   ``PlaylistBatch`` with ``PlaylistBatch.create``.
   Unsubscribing is the same as deleting a user playlist.